add strict typing to mikrotik
